conversions and tests

+++ src/rotary.c
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <wiringPi.h>
 #include <memory.h>
-#include "include/rotary.h"
+#include <rotary.h>
 
 #define DEGREES 360.00
 
@@ -81,11 +81,11 @@ int rotary_initialize(const struct config *config) {
     return 1;
 }
 
-double rotary_get_azimuth() {
+long double rotary_get_azimuth() {
     return azimuth_encoder.encoder_count * azimuth_encoder.degrees_per_encoded;
 }
 
-double rotary_get_altitude() {
+long double rotary_get_altitude() {
     return altitude_encoder.encoder_count * altitude_encoder.degrees_per_encoded;;
 }
